Little typo

+++ srcs/WebServer.cpp
@@ -60,7 +60,6 @@ void WebServer::run_select_poll(fd_set *reads, fd_set *writes)
 		std::cout << "[ERROR] select() timeout" << std::endl;
 	this->reads = *reads;
 	this->writes = *writes;
-
 }
 
 std::vector<int> WebServer::init_socket(std::map<int, t_server> server_list)
@@ -186,7 +185,7 @@ void	WebServer::reactor_loop(int epfd,std::map<int, t_server> server_list, std::
 
 void	WebServer::handle_client_request(struct epoll_event *current_event, int epfd, int i, std::map<int, t_server> server_list)
 {
-	size_t			ret = 0;
+	int			ret = 0;
 	std::cout << "\033[1m\033[35m \n Entering EPOLLIN and fd is "<< current_event[i].data.fd <<"\033[0m\n" << std::endl;
 	char buffer[30000] = {0};
